bring InputNumber2 into proper version 2 place
