--- docs/components/InputNumber2/InputNumber2.stories.mdx
@@ -1,22 +1,0 @@
-import { Canvas, Meta, Story } from "@storybook/addon-docs";
-import { InputNumber2 } from "@jobber/components/InputNumber2";
-import { Button } from "@jobber/components/Button";
-import { Content } from "@jobber/components/Content";
-
-<Meta
-  title="Components/Forms and Inputs/InputNumber2"
-  component={InputNumber2}
-/>
-
-# InputNumber2
-
-InputNumber2 is used in forms that accept numbers as an answer.
-
-## Design & usage guidelines
-
-This is best suited for input data that benefits from being modified in
-increments, such as quantity, price, or days (ie 2 days -> 3 days).
-
-While some types of data may technically be numbers, they can be ill-suited for
-using a number input. For example, phone numbers and credit card numbers provide
-no value to the user by offering an incrementer.
