+++ packages/components/src/InputNumber/InputNumber.rebuilt.module.css.d.ts
@@ -1,0 +1,21 @@
+declare const styles: {
+  readonly "container": string;
+  readonly "inline": string;
+  readonly "wrapper": string;
+  readonly "horizontalWrapper": string;
+  readonly "invalid": string;
+  readonly "disabled": string;
+  readonly "description": string;
+  readonly "fieldError": string;
+  readonly "small": string;
+  readonly "large": string;
+  readonly "center": string;
+  readonly "right": string;
+  readonly "inputWrapper": string;
+  readonly "childrenWrapper": string;
+  readonly "input": string;
+  readonly "label": string;
+  readonly "hideLabel": string;
+};
+export = styles;
+
