--- packages/components/src/InputNumber/index.ts
@@ -1,1 +1,0 @@
-export { InputNumber, InputNumberRef } from "./InputNumber";
