--- packages/components/src/InputNumber2/InputNumber2.module.css.d.ts
@@ -1,20 +1,0 @@
-declare const styles: {
-  readonly "center": string;
-  readonly "childrenWrapper": string;
-  readonly "container": string;
-  readonly "description": string;
-  readonly "disabled": string;
-  readonly "fieldError": string;
-  readonly "hideLabel": string;
-  readonly "horizontalWrapper": string;
-  readonly "inline": string;
-  readonly "input": string;
-  readonly "inputWrapper": string;
-  readonly "invalid": string;
-  readonly "label": string;
-  readonly "large": string;
-  readonly "right": string;
-  readonly "small": string;
-  readonly "wrapper": string;
-};
-export = styles;
