--- packages/components/src/InputNumber2/index.ts
@@ -1,1 +1,0 @@
-export { InputNumber2 } from "./InputNumber2";
