+++ packages/components/src/utils/meta/meta.json
@@ -111,7 +111,6 @@
     "InputFileContentContext.Provider",
     "InputGroup",
     "InputNumber",
-    "InputNumber2",
     "InputPassword",
     "InputPhoneNumber",
     "InputText",
