--- packages/site/src/content/InputNumber2/InputNumber2.props.json
@@ -1,52 +1,0 @@
-[
-  {
-    "tags": {},
-    "filePath": "../components/src/InputNumber2/InputNumber2.tsx",
-    "description": "",
-    "displayName": "InputNumber2",
-    "methods": [],
-    "props": {
-      "loud": {
-        "defaultValue": {
-          "value": false
-        },
-        "description": "Styles the text bold and uppercased",
-        "name": "loud",
-        "parent": {
-          "fileName": "../components/src/InputNumber2/InputNumber2.tsx",
-          "name": "InputNumber2Props"
-        },
-        "required": false,
-        "type": {
-          "name": "boolean"
-        }
-      },
-      "text": {
-        "defaultValue": null,
-        "description": "Text to display.",
-        "name": "text",
-        "parent": {
-          "fileName": "../components/src/InputNumber2/InputNumber2.tsx",
-          "name": "InputNumber2Props"
-        },
-        "required": true,
-        "type": {
-          "name": "string"
-        }
-      },
-      "onClick": {
-        "defaultValue": null,
-        "description": "Click handler.",
-        "name": "onClick",
-        "parent": {
-          "fileName": "../components/src/InputNumber2/InputNumber2.tsx",
-          "name": "InputNumber2Props"
-        },
-        "required": false,
-        "type": {
-          "name": "(event: React.MouseEvent<HTMLDivElement>) => void"
-        }
-      }
-    }
-  }
-]
