--- packages/site/src/content/InputNumber2/InputNumber2Notes.mdx
@@ -1,20 +1,0 @@
-{/* ## Configuration */}
-{/* - What are the composable parts of this component? */}
-{/* - What are the configuration options? */}
-
-{/* ## Component customization */}
-{/* - How can this component be customized? */}
-{/* - What are the customization options? */}
-
-{/* ## Controlled & uncontrolled */}
-{/* - How does this component behave when it is controlled or uncontrolled? */}
-{/* - How do you configure the component to be controlled or uncontrolled? */}
-
-{/* ## Platform considerations */}
-{/* - Are there any platform-specific considerations for this component? */}
-{/* - What kind of differences are there between iOS/Android or different browsers? */}
-
-{/* ## Developer notes */}
-{/* - What decisions worth remembering have gone into this component? */}
-{/* - Did we consider any potentially valuable functionality that we're not implementing right now for scope purposes? */}
-{/* - Are there any design decisions that might need explaining beyond the usage guidelines? */}
